ShowSubmissionPresenter with combined behavior
