+++ spec/presenters/show_submission_presenter_spec.rb
@@ -1,0 +1,6 @@
+require "./app/presenters/show_submission_presenter"
+require_relative "showing_a_submission_spec"
+
+describe ShowSubmissionPresenter do
+  it_behaves_like "showing a submission"
+end
